requirements.txt
arxiv
pypdf
aiohttp
python-dotenv
loguru
tqdm
apache-airflow
psycopg2-binary
sqlalchemy
pydantic
docling
tenacity
minio
opensearch-py
